1018: long double => double

+++ 1018/solve.cpp
@@ -20,8 +20,8 @@ using namespace std;
 
 typedef struct PointType
 {
-  long double x;
-  long double y;
+  double x;
+  double y;
 } Point;
 
 int relative_position(Point p, const vector<Point> &hill)
@@ -34,12 +34,12 @@ int relative_position(Point p, const vector<Point> &hill)
       break;
   }
 
-  long double x0 = hill[i].x;
-  long double y0 = hill[i].y;
-  long double x1 = hill[i-1].x;
-  long double y1 = hill[i-1].y;
+  double x0 = hill[i].x;
+  double y0 = hill[i].y;
+  double x1 = hill[i-1].x;
+  double y1 = hill[i-1].y;
 
-  long double hill_y = y1 - (y1-y0)*(x1-p.x)/(x1-x0);
+  double hill_y = y1 - (y1-y0)*(x1-p.x)/(x1-x0);
 
   if (fabsl(hill_y-p.y) < ACCURACY) {
     return 0;
@@ -49,22 +49,22 @@ int relative_position(Point p, const vector<Point> &hill)
   }
 }
 
-Point rotate_around(Point p, Point center, long double angle)
+Point rotate_around(Point p, Point center, double angle)
 {
-  long double dx = p.x - center.x;
-  long double dy = p.y - center.y;
-  long double l = sqrtl(dx * dx + dy * dy);
+  double dx = p.x - center.x;
+  double dy = p.y - center.y;
+  double l = sqrtl(dx * dx + dy * dy);
 
-  long double sin0 = dy / l;
-  long double cos0 = dx / l;
+  double sin0 = dy / l;
+  double cos0 = dx / l;
 
-  long double sin1 = cosl(angle)*sin0 + sinl(angle)*cos0;
-  long double cos1 = cosl(angle)*cos0 - sinl(angle)*sin0;
+  double sin1 = cosl(angle)*sin0 + sinl(angle)*cos0;
+  double cos1 = cosl(angle)*cos0 - sinl(angle)*sin0;
 
   dy = sin1 * l;
   dx = cos1 * l;
 
-  // printf("dx dy: %Lf %Lf\n", dx, dy);
+  // printf("dx dy: %lf %lf\n", dx, dy);
 
   p.x = center.x + dx;
   p.y = center.y + dy;
@@ -76,7 +76,7 @@ bool try_rotate(Point *center, Point *vertices, int n, const vector<Point> hill)
 {
   Point rotatable;
   int rotatable_index = -1;
-  long double initial_sign = 0.0;
+  double initial_sign = 0.0;
   int i;
   for (i = 0; i < n; ++i) {
     if (0 == relative_position(vertices[i], hill)) {
@@ -113,16 +113,16 @@ bool try_rotate(Point *center, Point *vertices, int n, const vector<Point> hill)
      }
   }
 
-  // printf("rotating center: %Lf, %Lf\n", rotatable.x, rotatable.y);
+  // printf("rotating center: %lf, %lf\n", rotatable.x, rotatable.y);
 
-  long double angle = 0.0 / M_PI;
-  long double step = 1.0 / M_PI / 180.0;
+  double angle = 0.0 / M_PI;
+  double step = 1.0 / M_PI / 180.0;
   bool decreasing_steps = false;
   bool new_contact_vertex = false;
 
   while (!new_contact_vertex) {
     angle += step * initial_sign;
-    // printf("angle: %Lf\n", angle);
+    // printf("angle: %lf\n", angle);
 
     bool over_rotated = false;
     for (i = 0; i < n; ++i) {
@@ -130,7 +130,7 @@ bool try_rotate(Point *center, Point *vertices, int n, const vector<Point> hill)
         continue;
 
       Point p = rotate_around(vertices[i], rotatable, angle);
-      // printf("%Lf %Lf\n", p.x, p.y);
+      // printf("%lf %lf\n", p.x, p.y);
 
       int comp = relative_position(p, hill);
       if (comp == 0) {
@@ -172,8 +172,8 @@ bool try_rotate(Point *center, Point *vertices, int n, const vector<Point> hill)
   center->y = p.y;
 
   // printf("index: %d\n", rotatable_index);
-  // printf("center: %Lf %Lf\n", center->x, center->y);
-  // printf("angle: %Lf\n", angle);
+  // printf("center: %lf %lf\n", center->x, center->y);
+  // printf("angle: %lf\n", angle);
 
   return true;
 }
@@ -181,17 +181,17 @@ bool try_rotate(Point *center, Point *vertices, int n, const vector<Point> hill)
 void calc_hill_articulation_points(
     vector<Point> &hill,
     Point start,
-    const vector<long double> &lengths,
-    const vector<long double> &slopes)
+    const vector<double> &lengths,
+    const vector<double> &slopes)
 {
   hill.push_back(start);
 
   int cnt = lengths.size();
   for (int i = 0; i < cnt; ++i) {
     Point p;
-    long double len = lengths[i];
-    long double slope = slopes[i];
-    long double angle = atanl(slope);
+    double len = lengths[i];
+    double slope = slopes[i];
+    double angle = atanl(slope);
     p.x = start.x - cosl(angle) * len;
     p.y = start.y - sinl(angle) * len;
 
@@ -213,28 +213,28 @@ void solve_case()
   vector<Point> hill;
 
   for (i = 0; i < n; ++i) {
-    scanf("%Lf %Lf", &(vertices[i].x), &(vertices[i].y));
+    scanf("%lf %lf", &(vertices[i].x), &(vertices[i].y));
   }
 
-  scanf("%Lf %Lf", &(center.x), &(center.y));
+  scanf("%lf %lf", &(center.x), &(center.y));
 
   Point line_start;
-  vector<long double> slopes;
-  vector<long double> lengths;
+  vector<double> slopes;
+  vector<double> lengths;
   bool finished = false;
 
   while (!finished) {
-    long double slope;
-    long double length;
+    double slope;
+    double length;
 
-    scanf("%Lf %Lf", &length, &slope);
+    scanf("%lf %lf", &length, &slope);
     slopes.push_back(slope);
     lengths.push_back(length);
 
     finished = slope == 0.0;
   }
 
-  scanf("%Lf %Lf", &(line_start.x), &(line_start.y));
+  scanf("%lf %lf", &(line_start.x), &(line_start.y));
 
   calc_hill_articulation_points(hill, line_start, lengths, slopes);
 
@@ -243,7 +243,7 @@ void solve_case()
   //try_rotate(&center, vertices, n, hill);
 
   Point p = center;
-  printf("%.3Lf %.3Lf\n", p.x, p.y);
+  printf("%.3lf %.3lf\n", p.x, p.y);
 }
 
 int main()
